refactor plugins
