+++ src/index.js
@@ -1,25 +1,13 @@
 import init from './init'
 import model from './model'
 import { getStore } from './store'
-import { pluginExports } from './plugins'
-import selectorsPlugin from '../src/plugins/selectors'
-import dispatchPlugin from '../src/plugins/dispatch'
-import effectsPlugin from '../src/plugins/effects'
-import hooksPlugin from '../src/plugins/hooks'
-
-const plugins = {
-  selectorsPlugin,
-  dispatchPlugin,
-  effectsPlugin,
-  hooksPlugin,
-}
+import { pluginExports } from './initPlugins'
 
 export default {
   init,
   model,
   getStore,
   pluginExports,
-  plugins
 }
 
 export {
@@ -27,5 +15,4 @@ export {
   model,
   getStore,
   pluginExports,
-  plugins
 }
